+++ src/tokens/themes/ember.css.d.ts
@@ -1,0 +1,3 @@
+declare const emberTheme: string;
+
+export default emberTheme;