+++ src/tokens/themes/lagoon.css.d.ts
@@ -1,0 +1,3 @@
+declare const lagoonTheme: string;
+
+export default lagoonTheme;